Move wayfinding.ts from lib/ -> client/
