+++ packages/shared/src/logic/index.ts
@@ -7,7 +7,6 @@ export * from './tiptap';
 export * from '@tloncorp/api/lib/hosting';
 export * from '@tloncorp/api/lib/noun';
 export * from './markdown';
-export * from '@tloncorp/api/lib/wayfinding';
 export * from '@tloncorp/api/lib/pinning';
 export * from '@tloncorp/api/client/activity';
 export * from '@tloncorp/api/client/branch';
@@ -16,3 +15,4 @@ export * from '@tloncorp/api/client/deeplinks';
 export * from '@tloncorp/api/client/postContent';
 export * from '@tloncorp/api/client/references';
 export * from '@tloncorp/api/client/utils';
+export * from '@tloncorp/api/client/wayfinding';
